pyproject.toml
[project]
name = "web-music-player"
version = "0.1.0"
description = "server"
readme = "README.md"
authors = [{ name = "Jason Swift", email = "anemele@outlook.com" }]
requires-python = ">=3.10"
dependencies = [
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
]

[tool.ruff]
lint.extend-select = ["F403", "I", "F401"]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
